README.md
The present work presents the development of an API for querying musical
data, in order to provide other developers a simplified access to a database with
information about songs and artists. This API was made in the Python programming
language, aided by the Pandas libraries. A graphical interface was built using
PySimpleGUI to present the results obtained through the API.